requirements.txt
alembic==1.10.4
astroid==2.4.2
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Login==0.6.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
isort==4.3.21
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==2.1.2
mccabe==0.6.1
pylint==2.5.3
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==2.0.12
toml==0.10.1
typing_extensions==4.5.0
Werkzeug==2.3.3
wrapt==1.12.1
WTForms==2.3.1
